+++ react-orbit/src/components/applications/ApplicationCard.tsx
@@ -1,5 +1,5 @@
 import { getTimeAgo } from "@/utils/date";
-import { Application } from "@/utils/localStorage";
+import { Application } from "@/utils/dataWrapper";
 import { getStatusStyles } from "@/utils/statusStyle";
 import { cn } from "@/lib/utils";
 import ApplicationActions from "./ApplicationActions";
