+++ react-orbit/src/pages/ApplicationsPage.tsx
@@ -7,7 +7,7 @@ import NewApplicationModal from "@/components/NewApplicationModal";
 import DeleteConfirmationModal from "@/components/applications/ConfirmDeleteModal";
 
 import { useApplications } from "@/hooks/useApplications";
-import { Application } from "@/utils/localStorage";
+import { Application } from "@/utils/dataWrapper";
 
 export function ApplicationsPage() {
   const [open, setOpen] = useState(false);
